A functional visualizer
City is drawn onto screen. There seems to be issue of using DISPLAY_SIZE as the width and height, where it causes an Index error, so for now I hard coded the width and the height. Also, the width and the height seems to also only allow the image to draw in that width and height section.